chore(CourseExecution): Move compensation tests to CreateCourseExecutionCompensationTest.groovy
